Move things around to try to break import cycles in go
