+++ src/glue-patterns/index.ts
@@ -1,0 +1,38 @@
+/**
+ * @module aws-glue-patterns
+ */
+
+/**
+  * @category Constructs
+  */
+export * from './alb-logs-table';
+
+/**
+  * @category Constructs
+  */
+export * from './cloudfront-logs-table';
+
+/**
+  * @category Constructs
+  */
+export * from './cloudtrail-table';
+
+/**
+  * @category Constructs
+  */
+export * from './flow-logs-table';
+
+/**
+  * @category Constructs
+  */
+export * from './s3-access-logs-table';
+
+/**
+  * @category Constructs
+  */
+export * from './ses-logs-table';
+
+/**
+  * @category Constructs
+  */
+export * from './waf-logs-table';
