+++ src/glue/jdbc-connection.ts
@@ -1,7 +1,7 @@
 import { ResourceProps, SecretValue } from 'aws-cdk-lib';
 import { ISecurityGroup, IVpc, SubnetSelection } from 'aws-cdk-lib/aws-ec2';
 import { Construct } from 'constructs';
-import { Connection, ConnectionType } from '../constructs/connection';
+import { Connection, ConnectionType } from '.';
 
 
 /**
