+++ src/glue/lib/glue-version.ts
