+++ src/index.ts
@@ -5,9 +5,10 @@
 export * as athena from './athena';
 export * as ec2 from './ec2';
 export * as glue from './glue';
+export * as glue_patterns from './glue-patterns';
 export * as kinesis_firehose from './kinesis-firehose';
 export * as ram from './ram';
-export * as s3 from './s3';
+export * as s3_patterns from './s3-patterns';
 export * as sso from './sso';
 
 export * as stacks from './stacks';
