+++ src/s3-patterns/cloudfront-logs-bucket.ts
@@ -1,13 +1,14 @@
 import { PhysicalName, ResourceProps } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
-import { Database, SesLogsTable } from '../../glue';
+import { Database } from '../glue';
+import { CloudfrontLogsTable } from '../glue-patterns';
 import { RawBucket } from './private/raw-bucket';
 
 
 /**
  * Configuration for objects bucket
  */
-export interface SesLogsBucketProps extends ResourceProps {
+export interface CloudfrontLogsBucketProps extends ResourceProps {
   readonly bucketName?: string;
   readonly createQueries?: boolean;
   readonly database?: Database;
@@ -15,10 +16,10 @@ export interface SesLogsBucketProps extends ResourceProps {
   readonly tableName?: string;
 }
 
-export class SesLogsBucket extends RawBucket {
+export class CloudfrontLogsBucket extends RawBucket {
   // Resource properties
   public readonly database: Database;
-  public readonly table: SesLogsTable;
+  public readonly table: CloudfrontLogsTable;
 
   // Input properties
   public readonly createQueries?: boolean;
@@ -26,14 +27,14 @@ export class SesLogsBucket extends RawBucket {
 
 
   /**
-     * Creates a new instance of the CloudtrailBucket class.
+     * Creates a new instance of the ElbLogsBucket class.
      *
      * @param scope A CDK Construct that will serve as this stack's parent in the construct tree.
      * @param id A name to be associated with the stack and used in resource naming. Must be unique
      * within the context of 'scope'.
      * @param props Arguments related to the configuration of the resource.
      */
-  constructor(scope: Construct, id: string, props: SesLogsBucketProps = {}) {
+  constructor(scope: Construct, id: string, props: CloudfrontLogsBucketProps = {}) {
     super(scope, id, {
       ...props,
       bucketEncryption: {
@@ -61,10 +62,10 @@ export class SesLogsBucket extends RawBucket {
     this.friendlyQueryNames = props.friendlyQueryNames;
 
     this.database = props.database ?? new Database(this, 'database', {
-      description: 'Database for storing SES event logs',
+      description: 'Database for storing ELB access logs',
     });
 
-    this.table = new SesLogsTable(this, 'table', {
+    this.table = new CloudfrontLogsTable(this, 'table', {
       bucket: this,
       createQueries: this.createQueries,
       database: this.database,
