+++ src/s3-patterns/index.ts
@@ -1,0 +1,40 @@
+/**
+ * @module aws-s3-patterns
+ */
+
+/**
+ * @category Constructs
+ */
+export * from './alb-logs-bucket';
+
+/**
+  * @category Constructs
+  */
+export * from './cloudfront-logs-bucket';
+
+/**
+  * @category Constructs
+  */
+export * from './cloudtrail-bucket';
+
+/**
+  * @category Constructs
+  */
+export * from './flow-logs-bucket';
+
+/**
+  * @category Constructs
+  */
+export * from './s3-access-logs-bucket';
+
+/**
+  * @category Constructs
+  */
+export * from './ses-logs-bucket';
+
+/**
+  * @category Constructs
+  */
+export * from './waf-logs-bucket';
+
+export * from './private/raw-bucket';
