--- src/s3/index.ts
@@ -1,40 +1,0 @@
-/**
- * @module aws-s3
- */
-
-/**
- * @category Patterns
- */
-export * from './patterns/alb-logs-bucket';
-
-/**
- * @category Patterns
- */
-export * from './patterns/cloudfront-logs-bucket';
-
-/**
- * @category Patterns
- */
-export * from './patterns/cloudtrail-bucket';
-
-/**
- * @category Patterns
- */
-export * from './patterns/flow-logs-bucket';
-
-/**
- * @category Patterns
- */
-export * from './patterns/s3-access-logs-bucket';
-
-/**
- * @category Patterns
- */
-export * from './patterns/ses-logs-bucket';
-
-/**
- * @category Patterns
- */
-export * from './patterns/waf-logs-bucket';
-
-export * from './patterns/private/raw-bucket';
